Unit Tests, Warnings and Clippy fixes
* Check if a process is running before dropping it, avoiding silent
  failing in unit tests.
* Add timeout for UDP tests.
* Fix warning on running tests regarding unused items in some features.
* Fix clippy issues.

+++ application/apps/indexer/sources/src/command/process.rs
@@ -38,8 +38,11 @@ pub struct ProcessSource {
 
 impl Drop for ProcessSource {
     fn drop(&mut self) {
-        if let Err(err) = self.process.start_kill() {
-            error!("Fail to kill child process: {err}");
+        let is_process_alive = self.process.try_wait().is_ok_and(|state| state.is_none());
+        if is_process_alive {
+            if let Err(err) = self.process.start_kill() {
+                warn!("Fail to kill child process: {err}");
+            }
         }
     }
 }
